Meta implation

+++ uninstall.inc.php
@@ -1,9 +1,19 @@
 <?php
 $REX['ADDON']['install']['asd_news'] = 0;
+$REX['ADDON']['installmsg']['asd_news']  = '';
 
-$sql = rex_sql::factory();
-$sql->setQuery('DROP TABLE IF EXISTS `' . $REX['TABLE_PREFIX'] . 'asd_news`');
-$sql->setQuery('DROP TABLE IF EXISTS `' . $REX['TABLE_PREFIX'] . 'asd_news_category`');
+
+include_once rex_path::addon('asd_news', 'classes/metainfo/rex_asd_metainfo_install.php');
+
+rex_asd_metainfo_install::setProperty();
+if($error = rex_asd_metainfo_install::delFields()) {
+    $REX['ADDON']['installmsg']['asd_news'] .= $error;
+    $REX['ADDON']['install']['asd_news'] = 1;
+} else {
+    $sql = rex_sql::factory();
+    $sql->setQuery('DROP TABLE IF EXISTS `' . $REX['TABLE_PREFIX'] . 'asd_news`');
+    $sql->setQuery('DROP TABLE IF EXISTS `' . $REX['TABLE_PREFIX'] . 'asd_news_category`');
+}
 
 if($sql->hasError()) {
     $msg = 'MySQL-Error: ' . $sql->getErrno() . '<br />';
@@ -13,12 +23,5 @@ if($sql->hasError()) {
     $REX['ADDON']['installmsg']['asd_news'] .= $msg;
 }
 
-include_once rex_path::addon('asd_news', 'classes/metainfo/rex_asd_metainfo_install.php');
-
-rex_asd_metainfo_install::setProperty();
-if($error = rex_asd_metainfo_install::delFields()) {
-    $REX['ADDON']['install']['asd_news'] .= $error;
-    $REX['ADDON']['install']['asd_news'] = 1;
-}
 
 ?>
